backend/requirements/base.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.0
django-cors-headers>=4.0
django-filter>=23.0
psycopg2-binary>=2.9
redis>=4.0
celery>=5.0
django-celery-beat>=2.0
cloudinary>=1.0
django-cloudinary-storage>=0.3
Pillow>=9.0
stripe>=5.0
WeasyPrint>=60.0
python-decouple>=3.0
django-unfold>=0.30
dj-database-url>=2.0
